made changes to movement will need to remake alot of the project. Still need to work on the actual movment so that the square follow each other.

+++ david_snake_JavaFX/src/main/java/marino/david_snake_javafx/Corner.java
@@ -1,11 +1,11 @@
 package marino.david_snake_javafx;
 
 public class Corner {
-    int x;
-    int y;
+    double x;
+    double y;
 
 
-    public Corner(int x, int y) {
+    public Corner(double x, double y) {
         this.x = x;
         this.y = y;
     }
